Refactor canvas view model to extract uml feature related functionalities
